api/requirements.txt
bentoml==1.3.21
catboost==1.2.7
joblib==1.4.2
pandas==2.0.3
pymysql==1.1.0
python-dotenv==1.0.1
scikit-learn==1.6.1
SQLAlchemy==1.4.54